tests/wpt/web-platform-tests/css/css-grid/parsing/grid-area-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Grid Layout Test: getComputedValue().gridArea</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#placement">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
// auto
test_computed_value("grid-area", "auto / auto / auto / auto");
test_computed_value("grid-row", "auto / auto");
test_computed_value("grid-column-end", "auto");

// <custom-ident>
test_computed_value("grid-row", "-zπ", "-zπ / -zπ");
test_computed_value("grid-row-start", "AZ");
test_computed_value("grid-column-start", "-_π");
test_computed_value("grid-row-end", "_9");


// <integer> && <custom-ident>?
test_computed_value("grid-area", "1 / 90 -a- / auto / auto");
test_computed_value("grid-row", "2 az / auto");
test_computed_value("grid-column", "9 / -19 zA");
test_computed_value("grid-row-start", "-19");
test_computed_value("grid-row-start", "9 -Z_");
test_computed_value("grid-column-start", "-44 Z");
test_computed_value("grid-row-end", "1 -πA");
test_computed_value("grid-column-end", "5 π_");

// span && [ <integer> || <custom-ident> ]
test_computed_value("grid-area", "span 2 i / auto / auto / auto");
test_computed_value("grid-row", "span 2 / auto");
test_computed_value("grid-column-start", "span 1 i");
test_computed_value("grid-row-end", "span 2 i");
test_computed_value("grid-column-end", "span 2");

// https://github.com/w3c/csswg-drafts/issues/3448
test_computed_value("grid-row-start", "span i");
test_computed_value("grid-row", "span i / auto");

// <grid-line> [ / <grid-line> ]{0,3}
test_computed_value("grid-area", "auto / i / auto / i");
test_computed_value("grid-area", "auto / i / 2 j", "auto / i / 2 j / i");
test_computed_value("grid-area", "auto / i / 2 j / span 3 k");
test_computed_value("grid-row", "auto / i");
test_computed_value("grid-column", "2 j / span 3 k");


// https://github.com/w3c/csswg-drafts/issues/2858
test_computed_value("grid-column-end",  "\\31st", "\\31 st");
test_computed_value("grid-column-end",  "\\31 st", "\\31 st");

</script>
</body>
</html>
